Change sample type to float

+++ include/types.h
@@ -1,6 +1,6 @@
 #ifndef __GLACIER_TYPES__
 #define __GLACIER_TYPES__
 
-typedef double SAMPLE;
+typedef float SAMPLE;
 
 #endif
